Minor changes on the Combat System & isInteracting bool
The PlayerCombatManager.cs on the Attack function now dynamically picks the animation to play , according to the current weapon on the right hand that the player has. Grabbing references of the current weapons from the InventoryManager.cs .

New scripts created:

InventoryManager.cs has a reference of the current right and left weapon.

WeaponHolderSlot.cs , with this script we can know in which hand the weapon is, letting up to new features like changing weapons.

WeaponSlotManager.cs handles all the references of the WeaponSlots , it also unloads, destroys and loads weapons on a specific slot.

PlayerHandler.cs will be the script that will connect all the other scripts in the future. For now gets the value of the isInteracting bool from Animator.

ResetIsInteracting.cs , a script for the animator, which as it name says, it resets the isInteracting bool to false when it exits the state of the current playing animation.

PlayerAnimatorManager.cs , it handles the function that will play all the animations receiving as parameters a string with the name of the animation, and a bool indicating if isInteracting or not.

The PlayerController.cs has been renamed as ---> PlayerMovement.cs

WeaponItem.cs now inherits from Item.cs which is the base Scriptable Object.